requirements.txt
Flask
opencv-python
numpy
onnxruntime
pytesseract
Pillow
gunicorn